requirements.txt
Django==3.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.3
redis==4.3.4
django-redis==5.2.0
